Populate user names in balanced :name_badge:
We were not passing up the name for buyers, and now we are.

Also it appears create_buyer may be deprecated in favor of Balanced::Account.new and Account#add_card according to the docs. We may want to consider moving towards that, although we're already fighting the heroku timeout as is with the one call