Attach merchant_uri to each debit object :nut_and_bolt:
Balanced has no way to associate debits with credits on their end, and they need that to do the fraud protection stuff for us.

This adds the column to our debit call. Mahmoud told me today in IRC that as long as we're passing it up like this, we should be good.

He also said this should be confirmable in the balanced dashboard soon.