Remove postal code for now :grey_exclamation:
We are getting a card validation error from Balanced when attempting to pass up a postal code. It claims we need a street address but this seems to be a problem on their end.

Waiting to hear back, in the mean time we are not passing postal code through to SM.Card