Store base views in a subfolder :open_file_folder:
Right now we only have BaseView and FormView, but we may well add CollectionView or other subclasses.

Working with JP today made me realize that it would be really helpful if each page was in the root, subviews are in subfolders, and the base views aren't cluttering up the main views.