Synchronize commitment statuses as they come in :dancers:
First we trigger a response on the plan when each ajax request to the commitments controller completes. Then we have the AdminParticipantsView listen for those and re-render.

We also had to modify the toJSON of AdminParticipantsView so that it was aware of the various states a commitment could be in, and the view conditionally displays success, failure, or a checkbox to charge the person.