Allow user to add a bank account after registration :bank:
We're using balanced js to create the bank account, so we had to set that up as well as allow a user to have a bank account uri set directly, rather than through bank_account_number and bank_routing_number.

We modified the callbacks on the user model, and added an SM.BankAccount class that functions almost identically to SM.Card

We also had to add a stub for the bank account, an SM.User.update method for adding the bank account uri to an existing account.

We augmented SM.FormView#value to look for the names of inputs as well as the js- ids.

Aside from that we set up the standard stuff for SM.PaymentInformationView.