Lay groundwork for converting a groupme access token into a splitme user :construction:
This introduces a UserConverter object, whose job it is to take a valid groupme access token and return a matching splitme user or a newly created splitme user with the groupme user data of name, email, and phone number.

It also introduces a supporting object GroupmeClient, that can make get requests to the groupme server given a token.