Allow groupme user to sign in to splitme :two_men_holding_hands:
This builds on the previous commit that allows us to exchange a groupme access token for a splitme one.

First we add an endpoint to employ the UserConverter object we created, /access_tokens/from_groupme_token.

Then we add a method SM.User.convert_groupme_token that hits /access_tokens/from_groupme_token and returns the user on success.

This method is then called when the user lands on /connect_with_groupme/:groupme_token. It is presumed that groupme web will have a link that looks like this for admins, and this shouldn't be hard to modify to accept a groupme email/phone and password, hitting groupme for the token and then splitme to convert.