Don't allow non-merchant user to charge any cards :no_good:
We had a hole in the system from the move to asynchronous bank account adding -- a user without a bank account was able to charge cards, which is not kosher for legal reasons.

This adds a check in the commitment object that verifies that the plan creator has a bank account before attempting a charge.