Don't allow a user to destroy their plan if it has been charged or collected :no_good:
We had a bug in our system where a user could destroy a plan that was partially or fully charged, or collected. As a result they could lose money to escrow or potentially destroy the record of a fradulent transaction.

This puts a check in a before_destroy on Plan to make sure it is not locked.