Allow user without bank account to add one during the flow of collecting a payment :potable_water:
We changed the system to not allow a user without a bank account to charge cards. This meant in the UI they were burnt if they didn't have a bank account.

This commit adds a check to the user before attempting a charge, and if the user has no bank account, it sends them to /payment_information, with a redirect set back to the plan.

In order to accomplish the redirect, we added a general utility object SM.Location that mirrors the functionality of authlogics store_location and redirect_back_or_default.